Initial commit
This bypasses mega's free 5GB limit per 6 hours using tor! You just need to enter the link in the file, and enjoy! I don't know whether it is ethical or not, I consider this as negligence of people at megga, which they have to figure out themselves! I just made this script because it is fun and I like it! and if it is possible, why not do it?